+++ SensateIoT.SmartEnergy.Dsmr.Api/App_Start/WebApiConfig.cs
@@ -10,12 +10,9 @@ namespace SensateIoT.SmartEnergy.Dsmr.Api
     {
         public static void Register(HttpConfiguration config)
         {
-	        var auth = GlobalConfiguration.Configuration.DependencyResolver.GetService(typeof(IAuthenticationRepository)) as IAuthenticationRepository;
-
 	        config.EnableCors();
 			config.MapHttpAttributeRoutes();
 			config.MessageHandlers.Add(new RequestLoggingMiddleware());
-			config.MessageHandlers.Add(new AuthenticationMiddleware(auth));
 
 			config.Formatters.JsonFormatter.SupportedMediaTypes
 				.Add(new MediaTypeHeaderValue("text/html"));