+++ SensateIoT.SmartEnergy.Dsmr.Api/Attributes/ExceptionHandlingAttribute.cs
@@ -3,13 +3,14 @@ using System.Net.Http;
 using System.Text;
 using System.Web.Http;
 using System.Web.Http.Filters;
+
 using log4net;
 using Newtonsoft.Json;
 
 using SensateIoT.SmartEnergy.Dsmr.Api.Data;
 using SensateIoT.SmartEnergy.Dsmr.Api.Exceptions;
 
-namespace SensateIoT.SmartEnergy.Dsmr.Api.Middleware
+namespace SensateIoT.SmartEnergy.Dsmr.Api.Attributes
 {
 	public class ExceptionHandlingAttribute : ExceptionFilterAttribute
 	{
@@ -17,12 +18,17 @@ namespace SensateIoT.SmartEnergy.Dsmr.Api.Middleware
 
 		public override void OnException(HttpActionExecutedContext context)
 		{
-			if(context.Exception is InvalidQueryArgumentException) {
+			switch(context.Exception) {
+			case HttpResponseException ex:
+				throw ex;
+			case InvalidQueryArgumentException _:
 				logger.Warn("Unable to complete request due to invalid query arguments.", context.Exception);
 				this.respondWithError("Unable to complete request with invalid query arguments.", context.Exception.Message, HttpStatusCode.BadRequest);
-			} else {
+				break;
+			default:
 				logger.Error("Unable to complete request!", context.Exception);
 				this.respondWithError("Unable to complete request.", null, HttpStatusCode.InternalServerError);
+				break;
 			}
 		}