+++ SensateIoT.SmartEnergy.Dsmr.Data/Models/Device.cs
@@ -5,7 +5,6 @@ namespace SensateIoT.SmartEnergy.Dsmr.Data.Models
 	public class Device
 	{
 		public int Id { get; set; }
-		public Guid OnboardingToken { get; set; }
 		public string ServiceName { get; set; }
 		public bool Enabled { get; set; }
 		public string PowerSensorId { get; set; }