+++ SensateIoT.SmartEnergy.Dsmr.DataAccess/Repositories/AuthenticationRepository.cs
@@ -1,16 +1,18 @@
 ﻿using System;
 using System.Data.SqlClient;
 using System.Linq;
+using System.Reflection.Emit;
 using System.Threading;
 using System.Threading.Tasks;
 
 using JetBrains.Annotations;
-
+using SensateIoT.SmartEnergy.Dsmr.Data.DTO;
 using SensateIoT.SmartEnergy.Dsmr.Data.Models;
 using SensateIoT.SmartEnergy.Dsmr.Data.Settings;
 using SensateIoT.SmartEnergy.Dsmr.DataAccess.Abstract;
 
 using Device = SensateIoT.SmartEnergy.Dsmr.Data.DTO.Device;
+using User = SensateIoT.SmartEnergy.Dsmr.Data.Models.User;
 
 namespace SensateIoT.SmartEnergy.Dsmr.DataAccess.Repositories
 {
@@ -20,6 +22,7 @@ namespace SensateIoT.SmartEnergy.Dsmr.DataAccess.Repositories
 		private const string DsmrApi_SelectProductToken = "DsmrApi_GetProductToken";
 		private const string DsmrApi_DsmrApi_SelectUserById = "DsmrApi_SelectUserById";
 		private const string DsmrApi_SelectDevicesForUser = "DsmrApi_SelectDevicesForUser";
+		private const string DsmrApi_Login = "DsmrApi_Login";
 
 		public AuthenticationRepository(AppSettings settings) : base(new SqlConnection(settings.DsmrProductConnectionString))
 		{
@@ -28,6 +31,12 @@ namespace SensateIoT.SmartEnergy.Dsmr.DataAccess.Repositories
 		public async Task<ProductToken> GetProductTokenAsync(Guid token, CancellationToken ct)
 		{
 			return await this.QuerySingleAsync<ProductToken>(DsmrApi_SelectProductToken, ct, "@token", token.ToString("D"))
+				.ConfigureAwait(false);
+		}
+
+		public async Task<LoginResult> LoginAsync(string email, string otp, CancellationToken ct)
+		{
+			return await this.QuerySingleAsync<LoginResult>(DsmrApi_Login, ct, "@email", email, "@token", otp)
 				.ConfigureAwait(false);
 		}