+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - 18-04-2021
+### Added
+- Authorization filter attributes
+- User login endpoint
+
+### Updated
+- All secure endpoints: added the authorization attribute
+
+### Removed
+- Authorization middleware
+- Request body logging
+
 ## [1.1.0] - 17-04-2021
 ### Added
 - Users controller